fix bugs; enable File field to use the same customProperty logic
